Basic READ.ME, improve asClickable on ItemProvider
Added the basic READ.ME structure for the project, incluiding code snippets for examples.

Only the code for utilities is missing, will be added on the next commit.

While changing the documentation I noticed ItemUtils should be renamed to ItemUtil, so it was done.

Finally, asClickable on ItemProvider was changed to a Consumer instead of the direct action, this was done so the developer does not need to use the CraftMenuAction directly.