improve sub-command node match algorithm
The matching alogrithm had some issues when the sub-command was empty or had arguments that were not inside the sub-command path.

We implemented an AlgorithmUtil for an improved result.
Now any sub-commands, even with extra arguments, should trigget its respective node.

An empty sub-command array still triggers an NPE, this will be fixed next commit before the version bump.