fix[spigot]: lib instances overriding
Since static instances are shared if on the same JVM, multiple plugins could not use tLib in the same server instance, due to JVM conflicts.

Now, plugins are required to save their tLib instance for later usage, replacing the old tLib.get().

This makes the development a little bit harder since you need to get the library from your plugin instance, but fixes the JVM problem which disabled the library completely for any plugin using it.

(A better solution may be implemented in the future)